requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
PyPDF2>=3.0.0
python-docx>=1.1.0
pandas>=2.0.0
openpyxl>=3.1.0
pdfplumber>=0.10.0
faiss-cpu>=1.7.0
requests>=2.31.0
numpy>=1.24.0
python-dotenv>=1.0.0
pdfplumber>=0.10.0